Ticket #10252 - Certfile/keyfile depend on mail sender.

+++ pyfred/modules/mailer.py
@@ -25,6 +25,9 @@ from email.MIMEText import MIMEText
 from email.Utils import formatdate, parseaddr
 # IMAP stuff
 import imaplib
+# certificate info
+import M2Crypto.X509, M2Crypto.m2
+from exceptions import Exception
 
 
 def convList2Array(list):
@@ -92,6 +95,95 @@ def filter_email_addrs(str):
             result += addr
     return result
 
+re_mail = re.compile("[-._a-zA-Z0-9]+@[-._a-zA-Z0-9]+")
+
+def smime_sender(certfile):
+    """
+    from certificate get email addresses of senders that can use the certificate
+    for email signing
+    """
+    sender = set()
+    try:
+        cert = M2Crypto.X509.load_cert(certfile)
+    except:
+        return sender
+    if not cert.check_purpose(M2Crypto.m2.X509_PURPOSE_SMIME_SIGN, 0):
+        return sender
+
+    # searches the subject name ...
+    try:
+        subj = cert.get_subject()
+        mails = subj.get_entries_by_nid(subj.nid['Email'])
+        for mail in mails:
+            sender.add(mail.get_data().as_text().lower())
+    except:
+        pass
+
+    # ... and the subject alternative name extension
+    try:
+        alt_name = cert.get_ext("subjectAltName").get_value()
+        mails = re_mail.findall(alt_name)
+        for mail in mails:
+            sender.add(mail.lower())
+    except:
+        pass
+
+    return sender
+
+
+def get_cert_key_pairs(conf, section, logger):
+    """
+    return array of cert/key pairs
+    """
+    cert = {}
+    for opt in conf.options(section):
+        if opt[:8]=="certfile":
+            name = opt[8:]
+            if not conf.has_option(section, "keyfile" + name):
+                raise Exception("certfile%s has not matching keyfile%s." % (name, name))
+            val = cert.get(name, {})
+            val['cert'] = conf.get(section, opt)
+            if not os.path.isfile(val['cert']):
+                raise Exception("%s=%s does not exist." % (opt, val['cert']))
+            cert[name] = val
+        elif opt[:7]=="keyfile":
+            name = opt[7:]
+            if not conf.has_option(section, "certfile" + name):
+                raise Exception("keyfile%s has not matching certfile%s." % (name, name))
+            val = cert.get(name, {})
+            val['key'] = conf.get(section, opt)
+            if not os.path.isfile(val['key']):
+                raise Exception("%s=%s does not exist." % (opt, val['key']))
+            cert[name] = val
+
+    pairs = []
+    for name in cert:
+        val = cert[name]
+        pairs.append({'cert':val['cert'], 'key':val['key']})
+        logger.log(logger.DEBUG, "Path to certfile%s is %s." % (name, val['cert']))
+        logger.log(logger.DEBUG, "Path to keyfile%s is %s." % (name, val['key']))
+    return pairs
+
+
+def get_sender_cert(conf, section, logger):
+    """
+    return dictionary where key is sender and value is array of cert/key pairs
+    that can be used for signing emails
+    """
+    sender_cert = {}
+    cert_key_pairs = get_cert_key_pairs(conf, section, logger)
+    for item in cert_key_pairs:
+        senders = smime_sender(item['cert'])
+        while 0<len(senders):
+            sender = senders.pop()
+            if sender_cert.has_key(sender):
+                sender_cert[sender].append(item)
+            else:
+                sender_cert[sender] = [item]
+            logger.log(logger.DEBUG, "Sender %s use certificate %s and key %s." % (sender, item['cert'], item['key']))
+    return sender_cert
+
+
 class Mailer_i (ccReg__POA.Mailer):
     """
     This class implements Mailer interface.
@@ -128,8 +220,7 @@ class Mailer_i (ccReg__POA.Mailer):
         self.idletreshold = 3600
         self.checkperiod = 60
         self.signing = False
-        self.keyfile = ""
-        self.certfile = ""
+        self.sender_cert = {}
         self.vcard = ""
         self.sendperiod = 300
         self.sendlimit = 100
@@ -209,23 +300,11 @@ class Mailer_i (ccReg__POA.Mailer):
                 self.signing = conf.getboolean("Mailer", "signing")
                 if self.signing:
                     self.l.log(self.l.DEBUG, "Signing of emails is turned on.")
-            except ConfigParser.NoOptionError, e:
-                pass
-            # certificate path
-            try:
-                certfile = conf.get("Mailer", "certfile")
-                if certfile:
-                    self.l.log(self.l.DEBUG, "Path to certfile is %s." %
-                            certfile)
-                    self.certfile = certfile
-            except ConfigParser.NoOptionError, e:
-                pass
-            # key path
-            try:
-                keyfile = conf.get("Mailer", "keyfile")
-                if keyfile:
-                    self.l.log(self.l.DEBUG, "Path to keyfile is %s." % keyfile)
-                    self.keyfile = keyfile
+                    # S/MIME certificates
+                    try:
+                        self.sender_cert = get_sender_cert(conf, "Mailer", self.l)
+                    except Exception, e:
+                        self.l.log(self.l.ERR, "get_sender_cert failure: %s" % e)
             except ConfigParser.NoOptionError, e:
                 pass
             # vcard switch
@@ -328,8 +407,8 @@ class Mailer_i (ccReg__POA.Mailer):
             cur = conn.cursor()
             cur.execute("SELECT mt.name "
                         "FROM mail_type mt "
-                        "WHERE (SELECT 1 FROM mail_type_default_map WHERE typeid=mt.id) IS NULL AND "
-                              "(SELECT 1 FROM mail_type_default_map WHERE typeid IS NULL) IS NULL")
+                        "WHERE (SELECT 1 FROM mail_type_mail_header_defaults_map WHERE typeid=mt.id) IS NULL AND "
+                              "(SELECT 1 FROM mail_type_mail_header_defaults_map WHERE typeid IS NULL) IS NULL")
             mailtype_without_default = cur.fetchall()
             cur.close()
             self.db.releaseConn(conn)
@@ -342,8 +421,8 @@ class Mailer_i (ccReg__POA.Mailer):
         if self.tester and not self.testmode:
             self.l.log(self.l.WARNING, "Tester configuration directive will "
                     "be ignored because testmode is not turned on.")
-        if self.signing and not (self.certfile and self.keyfile):
-            raise Exception("Certificate and key file must be set for mailer.")
+        if self.signing and not self.sender_cert:
+            raise Exception("Certificate and key file(s) must be set for mailer.")
         # do quick check that all files exist
         if not os.path.isfile(self.sendmail):
             raise Exception("sendmail binary (%s) does not exist." %
@@ -352,12 +431,6 @@ class Mailer_i (ccReg__POA.Mailer):
             if not os.path.isfile(self.openssl):
                 raise Exception("openssl binary (%s) does not exist." %
                         self.openssl)
-            if not os.path.isfile(self.certfile):
-                raise Exception("Certificate (%s) does not exist." %
-                        self.certfile)
-            if not os.path.isfile(self.keyfile):
-                raise Exception("Key file (%s) does not exist." %
-                        self.keyfile)
         # schedule regular cleanup
         joblist.append({ "callback":self.__search_cleaner, "context":None,
             "period":self.checkperiod, "ticks":1 })
@@ -564,7 +637,7 @@ class Mailer_i (ccReg__POA.Mailer):
         # get default values from database
         cur = conn.cursor()
         cur.execute("SELECT mhd.h_from,mhd.h_replyto,mhd.h_errorsto,mhd.h_organization,mhd.h_messageidserver "
-                    "FROM mail_type mt JOIN mail_type_default_map mtd ON (mtd.typeid=mt.id OR mtd.typeid IS NULL) "
+                    "FROM mail_type mt JOIN mail_type_mail_header_defaults_map mtd ON (mtd.typeid=mt.id OR mtd.typeid IS NULL) "
                     "LEFT JOIN mail_header_defaults mhd ON (mhd.id=mtd.defaultid) "
                     "WHERE mt.name=%s "
                     "ORDER BY mtd.typeid IS NULL "
@@ -880,6 +953,7 @@ class Mailer_i (ccReg__POA.Mailer):
         headers = mail[:headerend_index + 1]
         mimeheaders = ""
         signedmail = ""
+        sender = ""
         # throw away otherwise duplicated headers
         for header in headers.splitlines():
             if header.startswith("MIME-Version:") or \
@@ -888,14 +962,20 @@ class Mailer_i (ccReg__POA.Mailer):
                 mimeheaders += header + '\n'
             else:
                 signedmail += header + '\n'
+                if header.startswith("From:"):
+                    sender = re_mail.search(header[5:]).group(0).lower()
         mail = mimeheaders + mail[headerend_index + 1:]
+        if not self.sender_cert.has_key(sender):
+            self.l.log(self.l.ERR, "<%d> Sender %s has no S/MIME certificate." % (mailid, sender))
+            raise Mailer_i.MailerException("Signing of email failed.")
+        smime = self.sender_cert[sender][0]
         # create temporary file for openssl which will be used as input
         tmpfile = tempfile.mkstemp(prefix="pyfred-smime")
         os.write(tmpfile[0], mail)
         os.close(tmpfile[0])
         # do the signing
         stat, outdata, errdata = runCommand(mailid, "%s smime -sign -signer %s -inkey %s -in %s" %
-                   (self.openssl, self.certfile, self.keyfile, tmpfile[1]), None, self.l, retry_rounds=self.signing_cmd_retry_rounds)
+                   (self.openssl, smime['cert'], smime['key'], tmpfile[1]), None, self.l, retry_rounds=self.signing_cmd_retry_rounds)
         os.remove(tmpfile[1])
 
         if stat:
